// general-front/src/app/models/tecnico.ts
export interface Tecnico {
    id?:            any;
    firstName:      string;
    lastName:       string;
    cpf:            string;
    email:          string;
    address:        string;
    password:       string;
    perfis:         string[];
    gender:         string;
    enabled:        any;
    createdAt:      any;
  }
  